feat: 서버에서 영화 상세 모달을 html로 파싱하는 로직 추가
