+++ views/index.html
@@ -1,14 +1,14 @@
 <!DOCTYPE html>
 <html lang="ko">
-	<head>
-		<meta charset="UTF-8" />
-		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<title>영화 리뷰</title>
-		<link rel="stylesheet" href="./static/styles/index.css" />
-	</head>
-	<body>
-		<div id="root"></div>
-	</body>
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>영화 리뷰</title>
+    <link rel="stylesheet" href="/static/styles/index.css" />
+  </head>
+  <body>
+    <div id="root"></div>
+  </body>
 
-	<!--${INIT_DATA_AREA}-->
+  <!--${INIT_DATA_AREA}-->
 </html>
